pyproject.toml
[tool.poetry]
name = "orangewoodtest"
version = "0.1.0"
description = ""
authors = ["Keval Shah <keval.arhan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-genai = "^1.16.1"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
